\name{phenovisr-package}
\alias{phenovisr-package}
\alias{phenovisr}
\docType{package}
\title{
  Data-processing with phenological metrics
}
\description{
  The phenovisr package contains a set of
   data-processing functions to create histograms of
   phenological metrics (such as the green factor) for
   a set of pictures.
}
\details{
  Forthcoming.
}
\author{
Lucas Mello Schnorr

Maintainer: Lucas M. Schnorr <schnorr@inf.ufrgs.br>
}
\references{
  Forthcoming.
}
\keyword{ package }
\seealso{
  Forthcoming.
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
